plugin/grpc: Setup an async resource context before creating a span
This addresses GoogleCloudPlatform/cloud-trace-nodejs#656 by creating a new async context before launching the handler. Note that the `emitBefore` and `emitAfter` calls to `initialAsyncResource` are necessary for `cls-ah` to go back to a contex that existed before the call, or else it'll copy the same data and exhibit the same behaviour.